chore: add instill model with dummy execute
