+++ pkg/stabilityai/image_to_image_test.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -85,8 +84,4 @@ func TestGenerateImageFromImage(t *testing.T) {
 			assert.Equal(t, err, tt.expectedErr)
 		})
 	}
-}
-
-func TestTemp(t *testing.T) {
-	uuid.NewV4()
 }
